Fully working beta

+++ src/main/kotlin/repos/FilesRepo.kt
@@ -1,10 +1,9 @@
 package repos
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
+import GlobalState
 import models.BraveApp
 import models.DesktopFileData
+import java.awt.Desktop
 import java.io.File
 import java.nio.charset.Charset
 
@@ -13,18 +12,32 @@ import java.nio.charset.Charset
 // - all functions here should be pure
 object FilesRepo {
     
-    var files by mutableStateOf(readFiles())
-    
     private fun filterBraveFiles(file: File) : Boolean =
         file.name.startsWith("brave-") &&
             file.name.endsWith(".desktop")
     
     private fun parseLine(line: String) : String {
-        println("@debug parsing line: $line")
         val parts = line.split("=").toMutableList()
         parts.removeAt(0)
-        println("@debug line parts " + parts.joinToString("="))
         return parts.joinToString("=")
+    }
+    
+    private fun getTrashDirectory() : File {
+        val homeDir = File(System.getProperty("user.home"))
+        val trashDir = homeDir.resolve(".local/share/Trash")
+        val trashFilesDir = trashDir.resolve("files")
+        if (!trashDir.exists())
+            throw RuntimeException("Trash directory not found at $trashDir")
+        if (!trashFilesDir.exists()) {
+            println("Trash dir has no files subdirectory; creating... $trashFilesDir")
+            try {
+                trashFilesDir.mkdir()
+            } catch (e: Exception) {
+                e.printStackTrace()
+                throw RuntimeException("Failed to create trash files subdirectory at $trashFilesDir")
+            }
+        }
+        return trashFilesDir
     }
     
     private fun parseDesktopFile(lines: List<String>) : DesktopFileData {
@@ -39,7 +52,6 @@ object FilesRepo {
         StartupWMClass=crx_aghbiahbpaijignceidepookljebhfak
         OnlyShowIn=Old
         */
-        println("@debug parsing ${lines.size} lines")
         lines.forEach { 
             if (it.startsWith("Version"))
                 desktopData.Version = parseLine(it).toDoubleOrNull()
@@ -58,7 +70,6 @@ object FilesRepo {
             else if (it.startsWith("OnlyShowIn"))
                 desktopData.OnlyShowIn = parseLine(it)
         }
-        println("@debug desktopData $desktopData")
         return desktopData
     }
     
@@ -68,7 +79,7 @@ object FilesRepo {
         return BraveApp(file.name, file, details)
     }
     
-    fun readFiles() : List<BraveApp> {
+    fun readFiles() {
         
         val homeDir = File(System.getProperty("user.home"))
         val gnomeAppsDir = homeDir.resolve(".gnome/apps")
@@ -85,15 +96,68 @@ object FilesRepo {
             .filter(::filterBraveFiles)
             .forEach { braveFiles.add(getAppDetails(it)) }
         
-        return braveFiles.toList()
+        File("/tmp").walk().forEach {
+            braveFiles.add(
+                BraveApp(
+                    name = it.name,
+                    file = it,
+                    data = DesktopFileData()
+                )
+            )
+        }
+    
+        GlobalState.files = braveFiles.toList()
         
     }
     
-    fun deleteFile(file: File) {
+    fun getFilteredFiles() =
+        if (GlobalState.filterQuery.isEmpty()) GlobalState.files
+        else GlobalState.files.filter {
+            val name =
+                if (it.data.Name.isNullOrBlank()) it.name
+                else it.data.Name!!
+            name.lowercase().contains(
+                GlobalState.filterQuery.lowercase()
+            )
+        }
+    
+    fun deleteFile(file: File, permanent: Boolean = false) {
         
-        file.delete()
+        if (permanent) {
+            println("Permanently deleting ${file.name}")
+            file.delete()
+            return
+        }
         
-        files = readFiles()
+        try {
+            
+            // Try to delete the "correct" way first
+            println("Attempting to trash ${file.name}")
+            Desktop.getDesktop().moveToTrash(file)
+            
+        } catch (e: Exception) {
+            
+            try {
+                
+                val trashDir = getTrashDirectory()
+                val trashFile = trashDir.resolve(file.name)
+                
+                println("Attempting to manually trash ${file.name} to $trashDir")
+                file.copyTo(trashFile, overwrite = true)
+                file.delete()
+                
+                println("Trashed file ${file.name} to ${trashDir.absolutePath}")
+                
+            } catch (e: Exception) {
+    
+                e.printStackTrace()
+                println("Failed to trash ${file.name}")
+                
+                GlobalState.dialogFailedToDeleteAppOpen = true
+                
+            }
+            
+        }
         
     }
